requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==21.2.0
chardet==3.0.4
ChatterBot==1.0.4
chatterbot-corpus==1.2.0
click==8.0.1
discord-ext-menus==1.0.0a30+g309c702
discord.py==1.7.3
idna==3.2
joblib==1.0.1
mathparse==0.1.2
multidict==4.7.6
nltk==3.6.2
packaging==20.9
Pint==0.17
pymongo==3.11.4
pyparsing==2.4.7
python-dateutil==2.7.5
pytz==2021.1
PyYAML==3.13
regex==2021.4.4
six==1.16.0
SQLAlchemy==1.2.19
tqdm==4.61.1
typing-extensions==3.10.0.0
yarl==1.5.1